refactor: Refactor directory structures
