+++ src/main/kotlin/racingcar/domain/dtos/GameInput.kt
@@ -1,3 +1,3 @@
-package racingcar.dtos
+package racingcar.domain.dtos
 
 data class GameInput(val carNames: List<String>, val numberOfRounds: Int)
