+++ src/main/kotlin/racingcar/domain/dtos/GameResult.kt
@@ -1,3 +1,3 @@
-package racingcar.dtos
+package racingcar.domain.dtos
 
 data class GameResult(val roundResults: List<RoundResult>)
