+++ src/main/kotlin/racingcar/domain/dtos/RoundResult.kt
@@ -1,0 +1,5 @@
+package racingcar.domain.dtos
+
+import racingcar.domain.Car
+
+data class RoundResult(val cars: List<Car>)
