--- src/main/kotlin/racingcar/dtos/RoundResult.kt
@@ -1,5 +1,0 @@
-package racingcar.dtos
-
-import racingcar.Car
-
-data class RoundResult(val cars: List<Car>)
