--- src/main/kotlin/racingcar/inputviews/InputView.kt
@@ -1,7 +1,0 @@
-package racingcar.inputviews
-
-import racingcar.dtos.GameInput
-
-interface InputView {
-    fun receiveInput(): GameInput
-}
