--- src/main/kotlin/racingcar/resultviews/ResultView.kt
@@ -1,7 +1,0 @@
-package racingcar.resultviews
-
-import racingcar.dtos.GameResult
-
-interface ResultView {
-    fun printResult(gameResult: GameResult)
-}
