+++ src/main/kotlin/racingcar/views/inputviews/InputView.kt
@@ -1,0 +1,7 @@
+package racingcar.views.inputviews
+
+import racingcar.domain.dtos.GameInput
+
+interface InputView {
+    fun receiveInput(): GameInput
+}
