+++ src/main/kotlin/racingcar/views/resultviews/ResultView.kt
@@ -1,0 +1,7 @@
+package racingcar.views.resultviews
+
+import racingcar.domain.dtos.GameResult
+
+interface ResultView {
+    fun printResult(gameResult: GameResult)
+}
