+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,10 +1,11 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.domain.Car
 import kotlin.IllegalArgumentException
 
 class CarTest {
